refactor greedy algos
