+++ dttools/src/Makefile
@@ -19,6 +19,7 @@ SOURCES = \
 	bucketing_det_exhaust.c \
 	bucketing_det_greedy.c \
 	bucketing_exhaust_common.c \
+	bucketing_greedy_common.c \
 	bucketing_manager.c \
 	buffer.c \
 	catalog_query.c \
@@ -143,6 +144,7 @@ HEADERS_PUBLIC = \
 	bucketing_det_exhaust.h \
 	bucketing_det_greedy.h \
 	bucketing_exhaust_common.h \
+	bucketing_greedy_common.h \
 	bucketing_manager.h \
 	buffer.h \
 	category.h \
